requirements.txt
Django==1.11.22
djangorestframework==3.10.1
pytz==2019.1
